refactor(scripts): change files/folder arrangement in scripts
