+++ scripts/testnet/addRewardsAndBribes.ts
